+++ src/app/app.module.ts
@@ -1,10 +1,11 @@
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
+import { NgModule , APP_INITIALIZER} from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import { AppConfig } from './app.config';
 import { HeijunkaBoardComponent } from './heijunka-board/heijunka-board.component';
 import { ProjectComponent } from './project/project.component';
 import { ProjectAddComponent } from './project-add/project-add.component';
@@ -13,6 +14,10 @@ import { ProjectStateDetailsComponent } from './project-state-details/project-st
 import { KanbanCardCreateComponent } from './kanban-card-create/kanban-card-create.component';
 import { KanbanCardInProgressOverviewComponent } from './kanban-card-in-progress-overview/kanban-card-in-progress-overview.component';
 import { KanbanCardCompleteOverviewComponent } from './kanban-card-complete-overview/kanban-card-complete-overview.component';
+
+export function initializeApp(appConfig: AppConfig) {
+  return () => appConfig.load();
+}
 
 @NgModule({
   declarations: [
@@ -32,7 +37,10 @@ import { KanbanCardCompleteOverviewComponent } from './kanban-card-complete-over
     FormsModule,
     HttpClientModule
   ],
-  providers: [],
+  providers: [AppConfig,
+    { provide: APP_INITIALIZER,
+      useFactory: initializeApp,
+      deps: [AppConfig], multi: true }],
   bootstrap: [AppComponent]
 })
 export class AppModule { }